started refactor
