+++ src/components/RepoSearchBar/index.js
@@ -19,19 +19,16 @@ const RepoSearchBar = ({ setRepoQueryResults, allRepos }) => {
     }
 
     return (
-        <>
-            <form name="repo-search-bar" onSubmit={handleSubmit} className="py-4">
-                <div className="row">
-                    <div className="col">
-                        <input id="repo-search-bar" placeholder="repository name" className="border-bottom" onChange={handleUpdate} />
-                    </div>
-                    <div className="col">
-                        <input type="submit" value="submit" className="btn btn-secondary" />
-                    </div>
+        <form name="repo-search-bar" onSubmit={handleSubmit} className="py-4">
+            <div className="row">
+                <div className="col">
+                    <input id="repo-search-bar" placeholder="repository name" className="border-bottom" onChange={handleUpdate} />
                 </div>
-            </form >
-            {/* <hr className="mb-4" /> */}
-        </>
+                <div className="col">
+                    <input type="submit" value="submit" className="btn btn-secondary" />
+                </div>
+            </div>
+        </form >
     )
 }
