+++ src/components/index.js
@@ -1,5 +1,4 @@
-export { default as NavBar } from './NavBar';
-export { default as SearchForm } from './SearchForm';
 export { default as RepoCard } from './RepoCard';
 export { default as UserSearchBar } from './UserSearchBar';
 export { default as RepoSearchBar } from './RepoSearchBar';
+export { default as CardList } from '../components/CardList';
