+++ src/index.js
@@ -1,15 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { Provider } from 'react-redux';
-import { BrowserRouter as Router } from 'react-router-dom';
-import store from './store';
 import App from './App';
-import './styles/index.css';
 
-ReactDOM.render(
-    <Router>
-        <Provider store={store}>
-            <App />
-        </Provider>
-    </Router>,
-    document.getElementById("root"));
+ReactDOM.render(<App />, document.getElementById("root"));
