--- src/layout/index.js
@@ -1,1 +1,0 @@
-export { default as CardList } from './CardList';
